code review fixes
fix: code rabbit recommends
